Create GroupAnagrams.java
Given an array of strings strs, group the anagrams together. You can return the answer in any order.
Constraints:
1 <= strs.length <= 104
0 <= strs[i].length <= 100
strs[i] consists of lowercase English letters.

Solution:
Count the number of occurances of each letter for each string then group each string based on it's counts. 

Time complexity is O(m*n).
Space complexity is (m + p) where m is the number of strings, n is the average length of each string, and p is the number of unique anagrams.