Create TopKFrequentElements.java
Given an integer array nums and an integer k, return the k most frequent elements. You may return the answer in any order.
Constraints:
1 <= nums.length <= 105
-104 <= nums[i] <= 104
k is in the range [1, the number of unique elements in the array].
It is guaranteed that the answer is unique.

Solution:
Get the count for each unique int in nums. Put each unique num into a count array where each index is a list that represents a certain count. (index 3 is a list that contains the unique nums that have count 3)

Time complexity is O(n).
Space complexity is O(n).