Create LongestConsecutiveSequence.java
Given an unsorted array of integers nums, return the length of the longest consecutive elements sequence.
You must write an algorithm that runs in O(n) time.

Constraints:
0 <= nums.length <= 105
-109 <= nums[i] <= 109

Solution:
Add nums to hashset. Iterate through nums and check if num is the start of a sequence (num-1 not in sequence) then count how long the sequence is.

Time complexity is O(n).
Space complexity is O(n).