Create BestTimeToBuyAndSellStock.java
You are given an array prices where prices[i] is the price of a given stock on the ith day.
You want to maximize your profit by choosing a single day to buy one stock and choosing a different day in the future to sell that stock.
Return the maximum profit you can achieve from this transaction. If you cannot achieve any profit, return 0.

Constraints:
1 <= prices.length <= 105
0 <= prices[i] <= 104

Solution: Check if buy price is lower than sell price, if so then compare to maxprofit. If sell price is lower, use that as new buy price and continue checking sell prices.

Time complexity is O(n).
Space complexity is O(1).