Create PermutationInString.java
Given two strings s1 and s2, return true if s2 contains a permutation of s1, or false otherwise.
In other words, return true if one of s1's permutations is the substring of s2.

Constraints:
1 <= s1.length, s2.length <= 104
s1 and s2 consist of lowercase English letters.

Solution:
Create a hashmap for s1, and compare that hashmap to the hashmap of the current window of size s1.length. If they are the same then return true;

Time complexity is O(n).
Space complexity is O(26).