Create SlidingWindowMaximum.java
You are given an array of integers nums, there is a sliding window of size k which is moving from the very left of the array to the very right. You can only see the k numbers in the window. Each time the sliding window moves right by one position.
Return the max sliding window.

Constraints:
1 <= nums.length <= 105
-104 <= nums[i] <= 104
1 <= k <= nums.length

Solution:
Keep max value of window in the front of the deque. Instead of storing values, store indices so we can pop if front of queue is outside of window.

Time complexity is O(n).
Space complexity is O(n).