Create GenerateParentheses.java
Given n pairs of parentheses, write a function to generate all combinations of well-formed parentheses.

Constraints:
1 <= n <= 8

Solution:
Consider the number of open parentheses and the number of closed. There can never be closedCount > openCount in a valid combo so given this we use recursion to incrementally increase the number of open and closed parenthesis respectively.

Time complexity is based on n and the number of possible valid combinations.
Space complexity is also based on the number of possible valid combinations.