Create DailyTemperatures.java
Given an array of integers temperatures represents the daily temperatures, return an array answer such that answer[i] is the number of days you have to wait after the ith day to get a warmer temperature. If there is no future day for which this is possible, keep answer[i] == 0 instead.

Constraints:
1 <= temperatures.length <= 105
30 <= temperatures[i] <= 100

Solution:
Keep a monotonic decreasing stack paired with an index stack then iterate through the temperatures array. At each index, check if the temperature of the day is greater than the top of the stack. If it is then pop both the temperature stack and index stack then calculate the difference of days between the current index and the popped index, store in result array.

Time complexity is O(n).
Space complexity is O(n).