Create CarFleet.java
There are n cars going to the same destination along a one-lane road. The destination is target miles away.
You are given two integer array position and speed, both of length n, where position[i] is the position of the ith car and speed[i] is the speed of the ith car (in miles per hour).
A car can never pass another car ahead of it, but it can catch up to it and drive bumper to bumper at the same speed. The faster car will slow down to match the slower car's speed. The distance between these two cars is ignored (i.e., they are assumed to have the same position).
A car fleet is some non-empty set of cars driving at the same position and same speed. Note that a single car is also a car fleet.
If a car catches up to a car fleet right at the destination point, it will still be considered as one car fleet.
Return the number of car fleets that will arrive at the destination.

Constraints:
n == position.length == speed.length
1 <= n <= 105
0 < target <= 106
0 <= position[i] < target
All the values of position are unique.
0 < speed[i] <= 106

Solution:
Sort array. In a sorted array, we know the final position of each car will not change (there is no passing) therefore we can now focus only on arrival time. If the arrival time of a car would be lower or equal to the one ahead of it, that means that it would arrive at the same time as the one ahead of it (once again because there is no passing). Using this logic we can simply push to stack those cars that have only GREATER arrival times than the cars ahead which would indicate a new fleet. Return size of stack.

Time complexity is O(nlogn) because of sorting time but assuming a sorted array, the time complexity would be O(n).
Space complexity is O(n).